try to install git pre-commit hook only if .git folder exists (#2)
* try to install git pre-commit hook only if .git folder exists

* remove .idea folder from project gitignore

* added missing newline

+++ src/Composer/ScriptHandler.php
@@ -20,19 +20,20 @@ final class ScriptHandler
 
 	public static function addPhpCsToPreCommitHook()
 	{
-		$originFile = getcwd() . '/.git/hooks/pre-commit';
-		$templateContent = file_get_contents(__DIR__ . '/templates/git/hooks/pre-commit-phpcs');
-		if (file_exists($originFile)) {
-			$originContent = file_get_contents($originFile);
-			if (strpos($originContent, '# run phpcs') === FALSE) {
-				$newContent = $originContent . PHP_EOL . PHP_EOL . $templateContent;
-				file_put_contents($originFile, $newContent);
+		if (file_exists(getcwd() . '/.git')) {
+			$originFile = getcwd() . '/.git/hooks/pre-commit';
+			$templateContent = file_get_contents(__DIR__ . '/templates/git/hooks/pre-commit-phpcs');
+			if (file_exists($originFile)) {
+				$originContent = file_get_contents($originFile);
+				if (strpos($originContent, '# run phpcs') === FALSE) {
+					$newContent = $originContent . PHP_EOL . PHP_EOL . $templateContent;
+					file_put_contents($originFile, $newContent);
+				}
+			} else {
+				file_put_contents($originFile, $templateContent);
 			}
-
-		} else {
-			file_put_contents($originFile, $templateContent);
+			chmod($originFile, 0755);
 		}
-		chmod($originFile, 0755);
 	}
 
 }
